refactor: repository 도메인 패키지 하위로 변경
