+++ src/main/java/roomescape/reservation/domain/repository/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository;
+package roomescape.reservation.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
