+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -9,14 +9,14 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.reservation.dao.FakeReservationDao;
+import roomescape.reservation.dao.FakeReservationTimeDao;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.domain.repository.ReservationRepository;
+import roomescape.reservation.domain.repository.ReservationTimeRepository;
 import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.repository.FakeReservationDao;
-import roomescape.reservation.repository.FakeReservationTimeDao;
-import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation.repository.ReservationTimeRepository;
 
 @DisplayName("예약 로직 테스트")
 class ReservationServiceTest {
